make sure things are signed if they are expected to go negative
git-svn-id: https://svn.salilab.org/imp/trunk@8282 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/em2d/src/image_processing.cpp
@@ -499,7 +499,9 @@ void do_histogram_stretching(algebra::Matrix2D_d &m,
       break;
     }
   }
-  for (unsigned i=hist.size()-1;i>=0;--i ) {
+  // must be int so it can go negative
+  // and, unfortunately, unsigned - unsigned is still unsigned.
+  for (int i=static_cast<int>(hist.size())-1;i>=0;--i ) {
     if(hist[i] > h_cut) {
       i_max=i;
       break;
